Создание файлов util и data
