--- js/check-length-string.js
@@ -1,9 +1,0 @@
-// Функция для проверки максимальной длины строки
-function CheckLengthString(verifiedString, maxLength) {
-  if (verifiedString.length <= maxLength) {
-    return true;
-  }
-  return false;
-}
-
-export {CheckLengthString};
