+++ js/data.js
@@ -1,0 +1,10 @@
+import {getRandomIndexForId, getRandomIndexForUrl} from './util.js';
+
+export const MIN = 1;
+export const MAX = 25;
+export const INDEX_ID = getRandomIndexForId(MIN, MAX);
+export const INDEX_URL = getRandomIndexForUrl(MIN, MAX);
+export const MAX_COMMENT_LENGTH = 140;
+export const VERIFIED_STRING = 'message';
+export const PICTURE_COUNT = 25;
+
